+++ lib/arel/extensions.rb
@@ -20,6 +20,7 @@ require 'arel/extensions/lateral'
 require 'arel/extensions/range_function'
 require 'arel/extensions/with_ordinality'
 require 'arel/extensions/table'
+require 'arel/extensions/row'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
